Added displacement calculations for a rotating bar.
This calculates the displacement, curvature, shear, and moment for a beam rotating at rate omega. Since the beam is rotated, it is loaded with a centrifugal force. The values of interest are calculated using the Finite Element Method, employing the Hermite Shape Functions, and compared to the strong solution to check for convergence.